# docs/api/openapi-v2.yaml
openapi: 3.0.3
info:
  title: Flowora ERP API v2 M1
  version: 2.0.0-SNAPSHOT
  description: M1 contract for persistent identity, organization isolation, RBAC and master data.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
paths:
...
  /api/v2/session/login:
    post:
      operationId: loginV2
      security: []
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
        "423": { $ref: "#/components/responses/Error" }
...
  /api/v2/session/switch-organization:
    post:
      operationId: switchOrganization
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v2/session/change-password:
    post:
      operationId: changePassword
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v2/session/logout:
    post:
      operationId: logoutV2
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
...
  /api/v2/session/mfa/enroll:
    post:
      operationId: startMfaEnrollment
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /api/v2/session/mfa/confirm:
    post:
      operationId: confirmMfaEnrollment
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /api/v2/session/mfa/disable:
    post:
      operationId: disableMfa
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /api/v2/organizations:
    get:
      operationId: listOrganizations
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      operationId: createOrganization
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: SESSION }
  parameters:
    CsrfHeader: { name: X-XSRF-TOKEN, in: header, required: true, schema: { type: string } }
...
  responses:
    Success:
      description: Successful response with request correlation ID
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
    Error:
      description: Stable, localized API error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    ApiResponse:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string }
    ApiError:
      type: object
      required: [code, messageKey, args, fieldErrors, requestId]
      properties:
        code: { type: string }
        messageKey: { type: string }
        args: { type: object, additionalProperties: true }
        fieldErrors: { type: array, items: { $ref: "#/components/schemas/FieldError" } }
        requestId: { type: string }
    FieldError:
      type: object
      required: [field, code, messageKey]
      properties:
        field: { type: string }
        code: { type: string }
        messageKey: { type: string }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, maxLength: 190 }
        password: { type: string, format: password }
        mfaCode: { type: string, description: Six-digit TOTP or one-time recovery code. }